Improve inner container migration
